+++ src/Component/JsonSchema/Generator/NormalizerGenerator.php
@@ -6,6 +6,7 @@ use Jane\Component\JsonSchema\Generator\Context\Context;
 use Jane\Component\JsonSchema\Generator\Normalizer\DenormalizerGenerator;
 use Jane\Component\JsonSchema\Generator\Normalizer\JaneObjectNormalizerGenerator;
 use Jane\Component\JsonSchema\Generator\Normalizer\NormalizerGenerator as NormalizerGeneratorTrait;
+use Jane\Component\JsonSchema\Guesser\Guess\NonObjectGuessInterface;
 use Jane\Component\JsonSchema\Registry\Schema;
 use PhpParser\Comment;
 use PhpParser\Modifiers;
@@ -62,6 +63,10 @@ class NormalizerGenerator implements GeneratorInterface
         $normalizers = [];
 
         foreach ($schema->getClasses() as $class) {
+            if ($class instanceof NonObjectGuessInterface) {
+                continue;
+            }
+
             $modelFqdn = $schema->getNamespace() . '\\Model\\' . $class->getName();
 
             $methods = [];