+++ src/Component/OpenApi31/JaneOpenApi.php
@@ -2,6 +2,7 @@
 
 namespace Jane\Component\OpenApi31;
 
+use Jane\Component\JsonSchema\Generator\EnumGenerator;
 use Jane\Component\JsonSchema\Generator\Naming;
 use Jane\Component\JsonSchema\Generator\ValidatorGenerator;
 use Jane\Component\JsonSchema\JsonSchema\Normalizer\JsonSchemaNormalizer;
@@ -56,6 +57,9 @@ class JaneOpenApi extends CommonJaneOpenApi
         yield new RuntimeGenerator($naming, $parser);
         if ($options['validation'] ?? false) {
             yield new ValidatorGenerator($naming);
+        }
+        if ($options['enums-as-objects'] ?? false) {
+            yield new EnumGenerator();
         }
     }